Dużo pracy przy FormFaktura.cs i UCProdukt.cs
Generalnie w tym commicie wszystko się kręci wokół funkcjonalności
listy dostępnych produktów, które można dodać do faktury oraz
samego dodawania do faktury.

+++ FakturniakDataAccess/Models/ModelProdukt.cs
@@ -10,8 +10,8 @@ namespace FakturniakDataAccess.Models
     {
         public int id_produktu { get; set; }
         public string nazwa { get; set; }
-        public float cena_netto { get; set; }
-        public float cena_brutto { get; set; }
+        public Decimal cena_netto { get; set; }
+        public Decimal cena_brutto { get; set; }
         public int id_jednostki { get; set; }
         public int id_stawki { get; set; }
     }
